pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gira"
version = "0.1.0"
requires-python = ">=3.8"
description = "GIRA places JIRA tickets from your dependencies into your commit message"
readme = "README.md"
license.text = "MIT"
authors = [{name = "Tomas Peterka", email = "tomas.peterka@dronetag.cz"},]
dependencies = [
  "click",
  "pygit2",
  "tomli; python_version < '3.11'",
  "pyyaml",
  "jira",
]

[project.scripts]
gira = "gira.__main__:main"

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "wheel",
  "pre-commit",
]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["F", "E", "I"]  # F: pyflakes, E: pycodestyle, I: isort
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
